Home-made TCP and buffers refactoring
