--- src/main/java/sj/network/buffers/package-info.java
@@ -1,10 +1,0 @@
-/**
- * 
- */
-/**
- * Paquetage des données.
- * Va avec sj.network.tcp,
- * voir la documentation de ce package pour plus d'informations.
- * 
- */
-package sj.network.buffers;
