+++ src/main/java/sj/network/tcpAndBuffers/NetBufferTypeException.java
@@ -1,4 +1,4 @@
-package sj.network.buffers;
+package sj.network.tcpAndBuffers;
 
 public class NetBufferTypeException extends Exception {
 	private static final long serialVersionUID = -8785615402853344834L;
