+++ src/main/java/sj/network/tcpAndBuffers/TCPServer.java
@@ -1,4 +1,4 @@
-package sj.network.tcp;
+package sj.network.tcpAndBuffers;
 
 import java.net.InetAddress;
 import java.net.NetworkInterface;
@@ -7,8 +7,6 @@ import java.net.SocketException;
 import java.net.UnknownHostException;
 import java.util.ArrayList;
 import java.util.Enumeration;
-
-import sj.network.buffers.NetBuffer;
 
 /** Principe du TCPServer :
  * Un thread est créé pour accepter de nouveaux clients.
